It calculates the number of wet days for each month and each year

It calculates the number of wet days for each month and each year

nwetdays(data, valmin = 0.5, origin = "1961-1-1", station = names(data))

Arguments

data

data frame R object containing daily precipitation time series for several gauges (one gauge time series per column).

valmin

threshold precipitation value [mm] for wet/dry day indicator.

origin

character string "yyyy-mm-dd" indicated the date of the first row of "data".

station

character string indicating the stations. Default is names(data)

Value

Function returns a list of data frames containing the spell length expressed in days

Examples

 

data(trentino)

year_min <- 1961
year_max <- 1990

period <- PRECIPITATION$year>=year_min & PRECIPITATION$year<=year_max
station <- names(PRECIPITATION)[!(names(PRECIPITATION) %in% c("day","month","year"))]
prec_mes <- PRECIPITATION[period,station]  

## removing nonworking stations (e.g. time series with NA)
accepted <- array(TRUE,length(names(prec_mes)))
names(accepted) <- names(prec_mes)
for (it in names(prec_mes)) {
     accepted[it]  <- (length(which(!is.na(prec_mes[,it])))==length(prec_mes[,it]))
}

prec_mes <- prec_mes[,accepted]
## the dateset is reduced!!! 
prec_mes <- prec_mes[,1:3]

origin <- paste(year_min,1,1,sep="-")

nwetdays <- nwetdays(prec_mes,origin)
#>  chr [1:3] "T0001" "T0014" "T0021"
#> 'data.frame':	360 obs. of  5 variables:
#>  $ month: num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ year : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ T0001: num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ T0014: num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ T0021: num  0 0 0 0 0 0 0 0 0 0 ...