This function finds the bivariate joint probability or the binary correlation from the corresponding Gaussian correlation x

This function finds the bivariate joint probability or the binary correlation from the corresponding Gaussian correlation x

omega(x = 0.5, p0_v1 = 0.5, p0_v2 = NA, correlation = FALSE)

Arguments

x

value of expected correlation between the corresponding Gaussian-distributed variables

p0_v1, p0_v2

probability of no precipitation occurrences for the v1 and v2 time series respectively. See Notes.

correlation

logical numeric value. Default is FALSE. If TRUE the function returns the binary correlation like eq. 6 of Mhanna, et al.,2011.

Value

probability of no precipitation occurrence in both v1 and v2 simultaneously. It is a matrix if x is a matrix.

Note

This function makes use of normal copula. A graphical introduction to this function (with its inverse) makes is present in Mhanna and Bauwens (2011) and Wilks (1988) (See fig. 1 and par. 3.2) If the argument p0_v2, the two marginal probability values must be given as a vector through the argument p0_v1: p0_v1=c(p0_v1,p0_v2) . In case x is a correlation/covariance matrix the marginal probabilities are given as a vector through the argument p0_v1.

References

D.S. Wilks (1998), Multisite Generalization of a Daily Stochastic Precipitation Generation Model, Journal of Hydrology, Volume 210, Issues 1-4, September 1998, Pages 178-191, https://www.sciencedirect.com/science/article/pii/S0022169498001863

Muamaraldin Mhanna and Willy Bauwens (2011) A Stochastic Space-Time Model for the Generation of Daily Rainfall in the Gaza Strip, International Journal of Climatology, Volume 32, Issue 7, pages 1098-1112, doi:10.1002/joc.2305 , https://rmets.onlinelibrary.wiley.com/doi/10.1002/joc.2305

See also

Author

Emanuele Cordano

Examples

rho <- 0.4
p00 <- omega(x=rho,p0_v1=0.5,p0_v2=0.5)
cor00 <- omega(x=rho,p0_v1=0.5,p0_v2=0.5,correlation=TRUE)