This is the target function whose zero is searched to crete the inverse function of omega.

This is the target function whose zero is searched to crete the inverse function of omega.

omega_root(
  x = 0.5,
  p0_v1 = 0.5,
  p0_v2 = 0.5,
  p00 = p0_v1 * p0_v2,
  correlation = NA
)

Arguments

x

value of expected correlation between the corresponding Gaussian-distributed variables

p0_v1, p0_v2

probablity of no precipitatin occurrences for the v1 and v2 time series respectively.

p00

probability of no precipitation occurrence in both v1 and v2 simultanously returned by omega

correlation

numerical value. DEfault is NA. Binary correlation retured by omega when the argumet correlation=TRUE

Value

the value p00-omega(x=x,p0_v1=p0_v1,p0_v2=p0_v2) or correlation-omega(x=x,p0_v1=p0_v1,p0_v2=p0_v2) (if correlation is not NA)

Note

This function makes use of normal copula

See also

Author

Emanuele Cordano

Examples

rho <- 0.4
p00 <- omega(x=rho,p0_v1=0.5,p0_v2=0.5)
omega_root(x=rho,p0_v1=0.5,p0_v2=0.5,p00=p00)
#> [1] 0